Implement Day 2
Initially had a bug in part B where the first element wasn't correctly
skipped. Important to always think deeply about these edge cases. Might
take this opportunity to add tests to this repo for future problems, to
plan out testing for edge cases in advance. It would be cool to aim to
increase confidence in a successful first submission on future days.